% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERTModule.R
\name{BERT.message.device}
\alias{BERT.message.device}
\title{Create a generic message-based graphics device.}
\usage{
BERT.message.device(name, bgcolor = "white", width = 500, height = 350,
  pointsize = 14, type = "png")
}
\description{
Create a generic message-based graphics device.
}
