% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJ2XCLModule.R
\name{close.js.client.progress.bar}
\alias{close.js.client.progress.bar}
\title{Generic method for RJ2XCL shell progress bar}
\usage{
\method{close}{js.client.progress.bar}(pb)
}
\description{
Generic method for RJ2XCL shell progress bar
}
