% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJ2XCLModule.R
\name{history}
\alias{history}
\title{History implementation for the RJ2XCL console}
\usage{
history(max.show = 25, reverse = FALSE, pattern, ...)
}
\description{
History implementation for the RJ2XCL console
}
