% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R, R/package.module.R
\docType{package}
\name{module}
\alias{module}
\alias{module-package}
\title{Load and attach an R module in the current environment.}
\usage{
module(path)
}
\arguments{
\item{path}{character. If \code{path} is a character vector giving
an installed package in the current library path, a module
encompassing that package will be returned. Otherwise, the
directory given by the path will be loaded as a module.}
}
\value{
a \code{module} object.
}
\description{
Modules are light wrappers around environments that provide
better support for hierarching namespacing.

An R namespace system that supersedes packages by allowing nesting.
}

