% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODULEs.R
\name{MODULEs}
\alias{MODULEs}
\title{MODULEs function}
\usage{
MODULEs(edges, heatmap = TRUE)
}
\arguments{
\item{edges}{data.frame, two columns:interactorA,interactorB(for example 'PPI').}

\item{heatmap}{Default TRUE,heatmap of modules difference will plot if heatmap=FALSE}
}
\description{
Identify robust modules from a certain network
}
\author{
Fan Wang
}
