% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne.PCA.R
\name{ne.PCA}
\alias{ne.PCA}
\title{ne.PCA function}
\usage{
ne.PCA(nodes, edges, r = 0.8, core = 10)
}
\arguments{
\item{nodes}{data.frame, two columns:object,label(0:candidates/1:seeds).}

\item{edges}{data.frame, two columns:gene symbol A,gene symbol B.}

\item{r}{default 0.8,Restart probability of random walk,interval is (0,1).}

\item{core}{default 10,Top percentile of ranking for object,interval is (0,100].}
}
\description{
Node and edge Prioritization-based Community Analysis framework
}
\author{
Fan Wang <justin15751821901@gmail.com>
}
