% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERTModule.R
\name{pad.excel.array}
\alias{pad.excel.array}
\title{pad an excel array with empty strings}
\usage{
pad.excel.array(rslt)
}
\description{
if you enter an R function in an Excel array, but the function result
is smaller than the Excel range, the range will be filled with #N/A errors.
this is the correct behavior, but it can be unattractive. this function
will check the size of the output range, and if the data is smaller, will
fill the range with empty strings ("") that render as empty cells in Excel.
}
