% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.DEGs.R
\name{qc.DEGs}
\alias{qc.DEGs}
\title{qc.DEGs function}
\usage{
qc.DEGs(expr, group, lowEXP = 1, filter = 0.3, fc = 2, p = 0.05, func = FALSE)
}
\arguments{
\item{expr}{data.frame, Expression profile,the row name is the gene name,the column name is the sample name.}

\item{group}{data.frame, two columns:sample name, sample label,for example:GSM2718949,control.}

\item{lowEXP}{default 1,Low quality value cutoff,lowEXP must > 0.The expression value less than 'lowEXP' will be seen as missing values.}

\item{filter}{default 0.3,The proportion of the number of missing values in all samples in one row,the interval is (0,1).A row will be removed if the proportion of missing values in this row is greater than 'filter'.}

\item{fc}{default 2,Fold Change cutoff,fc must be greater than 1,generally fc is 1.5 or 2 or greater.}

\item{p}{default 0.05,Adjust p value cutoff,the interval is (0,1),generally p is 0.05 or 0.01 or 0.001.}

\item{func}{default FALSE,if 'func=TRUE',GO and KEGG pathways Enrichment analysis for DEGs(human genes) will be performed.}
}
\description{
Data pretreatment and differentially expressed genes identification
}
\author{
Fan Wang
}
