% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERTModule.R
\name{range.to.data.frame}
\alias{range.to.data.frame}
\title{Convert a list of lists into a data frame}
\usage{
\method{range}{to.data.frame}(rng, headers = F)
}
\description{
Spreadsheet functions that accept a range as an argument and range values 
returned by the COM/scripting API may be returned as list-of-list 
structures.  BERT does that to preserve type information; R vectors can't 
contain mixed types.  This function will convert a list-of-lists structure
into a data frame, optionally with column headers.
}
