% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJ2XCLModule.R
\name{RJ2XCL.console.graphics.device}
\alias{RJ2XCL.console.graphics.device}
\title{Create a console graphics device.}
\usage{
RJ2XCL.console.graphics.device(bgcolor = "white", width = 500, height = 350,
  pointsize = 14, type = "png")
}
\description{
\code{RJ2XCL.console.graphics.device} creates a graphics device that 
renders to the shell. type is either "svg" or "png".
}
