% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJ2XCLModule.R
\name{RJ2XCL.graphics.device}
\alias{RJ2XCL.graphics.device}
\title{Create a RJ2XCL graphics device.}
\usage{
RJ2XCL.graphics.device(name = "RJ2XCL-default", bgcolor = "white",
  width = 400, height = 400, pointsize = 12,
  scale = Sys.getenv("RJ2XCLGraphicsScale"), cell = F)
}
\description{
\code{RJ2XCL.graphics.device} creates a graphics device that renders to 
a named Shape in an Excel workbook.
}
\details{
If this is called from an Excel spreadsheet cell, set \code{cell=TRUE} 
and it will use a name that's generated from the cell address, which 
should be unique in the workbook.  (That won't survive sheet renaming, 
but in that case just toss the old one and a new one will be generated 
on the next paint).

Size, background and pointsize arguments are ignored if the target 
named shape already exists.  The values for these arguments are 
scaled based on reported screen DPI to give reasonable values on 
normal and high-DPI displays (to prevent this behavior, set the scale 
parameter to 1).
}
