% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCs.R
\name{TFCs}
\alias{TFCs}
\title{TFCs function}
\usage{
TFCs(edges)
}
\arguments{
\item{edges}{data.frame, two columns,for exmaple:gene symbol A, gene symbol B.}
}
\description{
Measure the role of interactions as an information medium
}
\author{
Fan Wang
}
