% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERTModule.R
\name{BERT.console.graphics.device}
\alias{BERT.console.graphics.device}
\title{Create a console graphics device.}
\usage{
BERT.console.graphics.device(bgcolor = "white", width = 400, height = 300,
  pointsize = 14, scale = Sys.getenv("BERTGraphicsScale"))
}
\description{
\code{BERT.console.graphics.device} creates a graphics device that 
renders to the shell.
}
