% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERTModule.R
\name{close.js.client.progress.bar}
\alias{close.js.client.progress.bar}
\title{Generic method for BERT shell progress bar}
\usage{
\method{close}{js.client.progress.bar}(pb)
}
\description{
Generic method for BERT shell progress bar
}
