% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\title{An S3 class representing metadata about a package.}
\usage{
package(name)
}
\arguments{
\item{name}{character. The name of the package. It
must be installed in the user's currently active
library.}
}
\description{
An S3 class representing metadata about a package.
}

